+++ src/props/binop.rs
@@ -337,13 +337,13 @@ mod tests {
     use dicetest::prelude::*;
     use std::collections::BTreeSet;
 
-    use crate::{elem, props, FateVarExt, Fun1, Fun2, Set};
+    use crate::{elem, props, Fun1, Fun2, Set};
 
     #[test]
     fn commutative_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("BTreeSet<u8>", dice::b_tree_set(dice::u8(..), ..));
-            let var = fate.roll_var(["x", "y"], set);
+            let var = fate.roll(set.var(["x", "y"]));
             let op = Fun2::new("intersection", |x, y| {
                 BTreeSet::<u8>::intersection(&x, &y)
                     .cloned()
@@ -357,7 +357,7 @@ mod tests {
     fn associative_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("Vec<u8>", dice::vec(dice::u8(..), ..));
-            let var = fate.roll_var(["x", "y", "z"], set);
+            let var = fate.roll(set.var(["x", "y", "z"]));
             let op = Fun2::new("append", |mut x, mut y| {
                 Vec::<u8>::append(&mut x, &mut y);
                 x
@@ -370,7 +370,7 @@ mod tests {
     fn left_distributive_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("i64", dice::i64(-1000..=1000));
-            let var = fate.roll_var(["x", "y", "z"], set);
+            let var = fate.roll(set.var(["x", "y", "z"]));
             let add = Fun2::infix("+", |x, y| x + y);
             let mul = Fun2::infix("*", |x, y| x * y);
             props::binop::left_distributive(var, add, mul);
@@ -381,7 +381,7 @@ mod tests {
     fn right_distributive_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("i64", dice::i64(-1000..=1000));
-            let var = fate.roll_var(["x", "y", "z"], set);
+            let var = fate.roll(set.var(["x", "y", "z"]));
             let add = Fun2::infix("+", |x, y| x + y);
             let mul = Fun2::infix("*", |x, y| x * y);
             props::binop::right_distributive(var, add, mul);
@@ -392,7 +392,7 @@ mod tests {
     fn distributive_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("i64", dice::i64(-1000..=1000));
-            let var = fate.roll_var(["x", "y", "z"], set);
+            let var = fate.roll(set.var(["x", "y", "z"]));
             let add = Fun2::infix("+", |x, y| x + y);
             let mul = Fun2::infix("*", |x, y| x * y);
             props::binop::distributive(var, add, mul);
@@ -403,7 +403,7 @@ mod tests {
     fn left_identity_elem_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("i8", dice::i8(..));
-            let var = fate.roll_var(["x"], set);
+            let var = fate.roll(set.var(["x"]));
             let op = Fun2::infix("+", |x, y| x + y);
             let e = elem("zero", 0);
             props::binop::left_identity_elem(var, op, e);
@@ -414,7 +414,7 @@ mod tests {
     fn right_identity_elem_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("i8", dice::i8(..));
-            let var = fate.roll_var(["x"], set);
+            let var = fate.roll(set.var(["x"]));
             let op = Fun2::infix("*", |x, y| x * y);
             let e = elem("one", 1);
             props::binop::right_identity_elem(var, op, e);
@@ -425,7 +425,7 @@ mod tests {
     fn identity_elem_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("f32", dice::f32(..));
-            let var = fate.roll_var(["x"], set);
+            let var = fate.roll(set.var(["x"]));
             let op = Fun2::infix("+", |x, y| x + y);
             let e = elem("zero", 0.0);
             props::binop::identity_elem(var, op, e);
@@ -436,7 +436,7 @@ mod tests {
     fn left_inverse_elem_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("i64", dice::i64(-1000..=1000));
-            let var = fate.roll_var(["x", "y"], set);
+            let var = fate.roll(set.var(["x", "y"]));
             let op = Fun2::infix("+", |x, y| x + y);
             let inv = Fun1::new("-", |x: i64| -x);
             props::binop::left_inverse_elem(var, op, inv);
@@ -447,7 +447,7 @@ mod tests {
     fn right_inverse_elem_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("i64", dice::i64(-1000..=1000));
-            let var = fate.roll_var(["x", "y"], set);
+            let var = fate.roll(set.var(["x", "y"]));
             let op = Fun2::infix("+", |x, y| x + y);
             let inv = Fun1::new("-", |x: i64| -x);
             props::binop::right_inverse_elem(var, op, inv);
@@ -458,7 +458,7 @@ mod tests {
     fn inverse_elem_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("i64", dice::i64(-1000..=1000));
-            let var = fate.roll_var(["x", "y"], set);
+            let var = fate.roll(set.var(["x", "y"]));
             let op = Fun2::infix("+", |x, y| x + y);
             let inv = Fun1::new("-", |x: i64| -x);
             props::binop::inverse_elem(var, op, inv);
@@ -469,7 +469,7 @@ mod tests {
     fn left_inverse_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("i64", dice::i64(-1000..=1000));
-            let var = fate.roll_var(["x", "y"], set);
+            let var = fate.roll(set.var(["x", "y"]));
             let op = Fun2::infix("+", |x, y| x + y);
             let invop = Fun2::infix("-", |x, y| x - y);
             props::binop::left_inverse(var, op, invop);
@@ -480,7 +480,7 @@ mod tests {
     fn right_inverse_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("i64", dice::i64(-1000..=1000));
-            let var = fate.roll_var(["x", "y"], set);
+            let var = fate.roll(set.var(["x", "y"]));
             let op = Fun2::infix("+", |x, y| x + y);
             let invop = Fun2::infix("-", |x, y| x - y);
             props::binop::right_inverse(var, op, invop);
@@ -491,7 +491,7 @@ mod tests {
     fn inverse_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("i64", dice::i64(-1000..=1000));
-            let var = fate.roll_var(["x", "y"], set);
+            let var = fate.roll(set.var(["x", "y"]));
             let op = Fun2::infix("+", |x, y| x + y);
             let invop = Fun2::infix("-", |x, y| x - y);
             props::binop::inverse(var, op, invop);
@@ -502,7 +502,7 @@ mod tests {
     fn equal_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("i64", dice::i64(-1000..=1000));
-            let var = fate.roll_var(["x", "y"], set);
+            let var = fate.roll(set.var(["x", "y"]));
             let op_1 = Fun2::new("add", |x, y| x + y);
             let op_2 = Fun2::new("add_assign", |mut x, y| {
                 x += y;