+++ src/props/binrel.rs
@@ -230,13 +230,13 @@ where
 mod tests {
     use dicetest::prelude::*;
 
-    use crate::{props, FateVarExt, Fun2, Set};
+    use crate::{props, Fun2, Set};
 
     #[test]
     fn reflexive_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("u8", dice::u8(..));
-            let var = fate.roll_var(["x"], set);
+            let var = fate.roll(set.var(["x"]));
             let rel = Fun2::infix("==", |x, y| x == y);
             props::binrel::reflexive(var, rel);
         })
@@ -246,7 +246,7 @@ mod tests {
     fn symmetric_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("f32", dice::any_f32());
-            let var = fate.roll_var(["x", "y"], set);
+            let var = fate.roll(set.var(["x", "y"]));
             let rel = Fun2::infix("!=", |x, y| x != y);
             props::binrel::symmetric(var, rel);
         })
@@ -256,7 +256,7 @@ mod tests {
     fn asymmetric_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("f32", dice::any_f32());
-            let var = fate.roll_var(["x", "y"], set);
+            let var = fate.roll(set.var(["x", "y"]));
             let rel = Fun2::infix("<", |x, y| x < y);
             props::binrel::asymmetric(var, rel);
         })
@@ -266,7 +266,7 @@ mod tests {
     fn antisymmetric_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("u8", dice::u8(..));
-            let var = fate.roll_var(["x", "y"], set);
+            let var = fate.roll(set.var(["x", "y"]));
             let rel = Fun2::infix("<", |x, y| x < y);
             props::binrel::antisymmetric(var, rel);
         })
@@ -276,7 +276,7 @@ mod tests {
     fn connex_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("u8", dice::u8(..));
-            let var = fate.roll_var(["x", "y"], set);
+            let var = fate.roll(set.var(["x", "y"]));
             let rel = Fun2::infix("<=", |x, y| x <= y);
             props::binrel::connex(var, rel);
         })
@@ -286,7 +286,7 @@ mod tests {
     fn transitive_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("char", dice::char());
-            let var = fate.roll_var(["x", "y", "z"], set);
+            let var = fate.roll(set.var(["x", "y", "z"]));
             let rel = Fun2::infix("<", |x, y| x < y);
             props::binrel::transitive(var, rel);
         })
@@ -296,7 +296,7 @@ mod tests {
     fn partial_equivalence_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("f32", dice::any_f32());
-            let var = fate.roll_var(["x", "y", "z"], set);
+            let var = fate.roll(set.var(["x", "y", "z"]));
             let rel = Fun2::infix("==", |x, y| x == y);
             props::binrel::partial_equivalence(var, rel);
         })
@@ -306,7 +306,7 @@ mod tests {
     fn equivalence_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("String", dice::string(dice::char(), ..));
-            let var = fate.roll_var(["x", "y", "z"], set);
+            let var = fate.roll(set.var(["x", "y", "z"]));
             let rel = Fun2::infix("==", |x, y| x == y);
             props::binrel::equivalence(var, rel);
         })
@@ -316,7 +316,7 @@ mod tests {
     fn partial_order_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("u8²", dice::zip().two(dice::u8(..), dice::u8(..)));
-            let var = fate.roll_var(["x", "y", "z"], set);
+            let var = fate.roll(set.var(["x", "y", "z"]));
             let rel = Fun2::infix("<=", |x: (u8, u8), y: (u8, u8)| {
                 (x.0 <= y.0) && (x.1 <= y.1)
             });
@@ -328,7 +328,7 @@ mod tests {
     fn total_order_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("u8", dice::u8(..));
-            let var = fate.roll_var(["x", "y", "z"], set);
+            let var = fate.roll(set.var(["x", "y", "z"]));
             let rel = Fun2::infix("<=", |x, y| x <= y);
             props::binrel::total_order(var, rel);
         })