+++ src/props/fun.rs
@@ -160,7 +160,7 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::{props, FateVarExt, Fun1, Fun2, Set};
+    use crate::{props, Fun1, Fun2, Set};
 
     use dicetest::prelude::*;
     use std::collections::BTreeSet;
@@ -176,7 +176,7 @@ mod tests {
                     ..,
                 ),
             );
-            let var = fate.roll_var(["x"], set);
+            let var = fate.roll(set.var(["x"]));
             let f = Fun1::new("trim", |x: String| x.trim().to_owned());
             props::fun::idempotent(var, f);
         })
@@ -186,7 +186,7 @@ mod tests {
     fn left_inverse_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("f32", dice::f32(..));
-            let var = fate.roll_var(["x"], set);
+            let var = fate.roll(set.var(["x"]));
             let f = Fun1::new("to_string", |x: f32| x.to_string());
             let g = Fun1::new("from_string", |y: String| f32::from_str(&y).unwrap());
             props::fun::left_inverse(var, f, g);
@@ -197,7 +197,7 @@ mod tests {
     fn right_inverse_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("i8", dice::i8(..));
-            let var = fate.roll_var(["x"], set);
+            let var = fate.roll(set.var(["x"]));
             let f = Fun1::new("from_string", |x: String| i8::from_str(&x).unwrap());
             let g = Fun1::new("to_string", |y: i8| y.to_string());
             props::fun::right_inverse(var, f, g);
@@ -209,8 +209,8 @@ mod tests {
         Dicetest::once().run(|mut fate| {
             let set_s = Set::new("i8", dice::i8(..));
             let set_t = Set::new("u8", dice::u8(..));
-            let var_s = fate.roll_var(["x"], set_s);
-            let var_t = fate.roll_var(["y"], set_t);
+            let var_s = fate.roll(set_s.var(["x"]));
+            let var_t = fate.roll(set_t.var(["y"]));
             let f = Fun1::new("from_string", |x: i8| x as u8);
             let g = Fun1::new("to_string", |y: u8| y as i8);
             props::fun::inverse(var_s, var_t, f, g);
@@ -221,7 +221,7 @@ mod tests {
     fn equal_1_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("u64", dice::u64(..1000));
-            let var = fate.roll_var(["x"], set);
+            let var = fate.roll(set.var(["x"]));
             let f = Fun1::postfix("+2", |x: u64| x + 2);
             let g = Fun1::postfix("+1+1", |x: u64| x + 1 + 1);
             props::fun::equal_1(var, f, g);
@@ -233,8 +233,8 @@ mod tests {
         Dicetest::once().run(|mut fate| {
             let set_s = Set::new("BTreeSet<u8>", dice::b_tree_set(dice::u8(..), ..));
             let set_t = Set::new("u8", dice::u8(..));
-            let var_s = fate.roll_var(["xs"], set_s);
-            let var_t = fate.roll_var(["x"], set_t);
+            let var_s = fate.roll(set_s.var(["xs"]));
+            let var_t = fate.roll(set_t.var(["x"]));
             let f = Fun2::new("insert", |mut xs: BTreeSet<u8>, x| {
                 xs.insert(x);
                 xs
@@ -252,7 +252,7 @@ mod tests {
     fn commutative_example() {
         Dicetest::once().run(|mut fate| {
             let set = Set::new("BTreeSet<u8>", dice::b_tree_set(dice::u8(..), ..));
-            let var = fate.roll_var(["x", "y"], set);
+            let var = fate.roll(set.var(["x", "y"]));
             let f = Fun2::new("is_disjoin", |x, y| BTreeSet::<u8>::is_disjoint(&x, &y));
             props::fun::commutative(var, f);
         })