+++ src/var.rs
@@ -1,10 +1,8 @@
 use dicetest::hint;
-use dicetest::{dice, Die, DieOnce, Fate};
-use std::array::IntoIter;
 use std::fmt::Debug;
 use std::fmt::Write;
 
-use crate::{elem, Elem, Eval, Set};
+use crate::{Elem, Eval};
 
 /// Represents elements that were chosen from a set and that can be used as variables in a
 /// predicate.
@@ -50,52 +48,5 @@ impl<'a, S: Debug, const N: usize> Var<'a, S, N> {
     pub fn as_ref<'b: 'a>(&'b self) -> Var<'a, &'b S, N> {
         let elems = array_init::array_init(|i| self.elems[i].as_ref());
         Var::new(self.set, elems)
-    }
-}
-
-pub trait Sealed {}
-
-/// Extension for [`Fate`] that allows to choose variables randomly from a set represented
-/// by a generator.
-pub trait FateVarExt: Sealed {
-    /// Returns a [`Var`] with a single value generated with the given [`DieOnce`].
-    fn roll_single_var<'a, S: Debug, D: DieOnce<S>>(
-        &mut self,
-        name: &'a str,
-        set: Set<'a, S, D>,
-    ) -> Var<'a, S, 1>;
-
-    /// Returns a [`Var`] with `N` values generated with the given [`Die`].
-    fn roll_var<'a, S: Debug, D: Die<S>, const N: usize>(
-        &mut self,
-        names: [&'a str; N],
-        set: Set<'a, S, D>,
-    ) -> Var<'a, S, N>;
-}
-
-impl Sealed for Fate<'_> {}
-
-impl FateVarExt for Fate<'_> {
-    fn roll_single_var<'a, S: Debug, D: DieOnce<S>>(
-        &mut self,
-        name: &'a str,
-        set: Set<'a, S, D>,
-    ) -> Var<'a, S, 1> {
-        let value = self.roll(set.elem_die);
-        let elem = elem(name, value);
-        Var::new(set.name, [elem])
-    }
-
-    fn roll_var<'a, S: Debug, D: Die<S>, const N: usize>(
-        &mut self,
-        names: [&'a str; N],
-        set: Set<'a, S, D>,
-    ) -> Var<'a, S, N> {
-        let values = self.roll(dice::array::<_, _, N>(set.elem_die));
-        let elems_iter = IntoIter::new(names)
-            .zip(IntoIter::new(values))
-            .map(|(name, value)| elem(name, value));
-        let elems: [_; N] = array_init::from_iter(elems_iter).unwrap();
-        Var::new(set.name, elems)
     }
 }