+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add struct `diceprop::Var`. It is implemented with const generics and can represent an arbitrary number of variables. It replaces `diceprop::{Var1, Var2, Var3}`.
 - Add struct `diceprop::Set`. It wraps a value generator and is necessary for generating `diceprop::Var`.
 - The functions `diceprop::{fun1, postfix_fun1, fun2, infix_fun2}` are now associated functions of `diceprop::{Fun1, Fun2}`.
+- Remove `diceprop::FateVarExt`. Use the generators `diceprop::Set::{single_var, var}` instead.
 
 ### Added
 - Add function `diceprop::props::binop::equal`